+++ src/main/kotlin/ru/meanmail/codeInspection/InstalledPackageInspection.kt
@@ -1,6 +1,5 @@
 package ru.meanmail.codeInspection
 
-import com.intellij.codeInspection.LocalInspectionTool
 import com.intellij.codeInspection.LocalInspectionToolSession
 import com.intellij.codeInspection.ProblemHighlightType
 import com.intellij.codeInspection.ProblemsHolder